Keep user input when edit tag (with validation error occurs)
It is more usable to keep input of the form when validation error occurs.

If keep `name`, it chenges the action URL of the form.
I guess this is the reason why input is now discarded.

Therefore, separate `@tag` and `@tag_for_input`.
`@tag` holds the tag before changed.
`@tag_for_input` accepts user input.